+++ ai-service/src/main/java/com/ojosama/AiServiceApplication.java
@@ -1,4 +1,4 @@
-package com.ojosama.aiservice;
+package com.ojosama;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
