--- ai-service/src/main/java/com/ojosama/aiservice/domain/ModerationResult.java
@@ -1,9 +1,0 @@
-package com.ojosama.aiservice.domain;
-
-import com.ojosama.operationservice.domain.model.enums.ReportCategory;
-
-public record ModerationResult(
-        boolean isViolating,
-        ReportCategory category,
-        String reason
-) { }
